Move TranslationLanguagePickerResultsController to its own file
